moved add_star to journalist_app/col.py
