+++ securedrop/journalist_app/__init__.py
@@ -12,7 +12,7 @@ import template_filters
 import version
 
 from db import db_session, Journalist
-from journalist_app import account, admin, main
+from journalist_app import account, admin, main, col
 from journalist_app.utils import get_source
 
 
@@ -88,5 +88,6 @@ def create_app(config):
     app.register_blueprint(account.make_blueprint(config),
                            url_prefix='/account')
     app.register_blueprint(admin.make_blueprint(config), url_prefix='/admin')
+    app.register_blueprint(col.make_blueprint(config), url_prefix='/col')
 
     return app
