+++ securedrop/journalist_templates/_source_row.html
@@ -8,7 +8,8 @@
       <input id="checkbox" type="checkbox" name="cols_selected" value="{{ source.filesystem_id }}">
       <span class="code-name"><a href="/col/{{ source.filesystem_id }}" id="starred-source-link-{{ loop_index }}">{{ source.journalist_designation }}</a></span>
     {% else %}
-      <button class="button-star un-starred" type="submit" formaction="/col/add_star/{{ source.filesystem_id }}"><i class="fa fa-star"></i></button>
+      <button class="button-star un-starred" type="submit"
+        formaction="{{ url_for('col.add_star', filesystem_id=source.filesystem_id) }}"><i class="fa fa-star"></i></button>
       <input id="checkbox" type="checkbox" name="cols_selected" value="{{ source.filesystem_id }}">
       <span class="code-name"><a href="/col/{{ source.filesystem_id }}" id="un-starred-source-link-{{ loop_index }}">{{ source.journalist_designation }}</a></span>
     {% endif %}
